feat(QueryHistory): 컬럼 구조 조정
