--- backend/src/main/java/com/gigigenie/domain/product/entity/QueryHistory.java
@@ -1,45 +1,0 @@
-package com.gigigenie.domain.product.entity;
-
-import com.gigigenie.domain.member.entity.Member;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.*;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-@Setter
-@Entity
-@Table(name = "query_history")
-public class QueryHistory {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "query_id", nullable = false)
-    private Integer id;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    @Size(max = 255)
-    @NotNull
-    @Column(name = "query_text", nullable = false)
-    private String queryText;
-
-    @NotNull
-    @Column(name = "response_text", nullable = false, length = Integer.MAX_VALUE)
-    private String responseText;
-
-    @NotNull
-    @Column(name = "query_time", nullable = false)
-    private Long queryTime;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-}
